added useFormDirty
